+++ lib/nostr_dart.dart
@@ -3,4 +3,4 @@ library nostr;
 export 'src/nostr.dart';
 export 'src/keys.dart';
 export 'src/event.dart';
-export 'src/contact_list.dart';
+export 'src/contact_list.dart' show Contact, ContactList;