+++ CHANGELOG.md
@@ -4,8 +4,13 @@
 
 ## 0.1.1
 
-- Fix package import documentation.
+- Fixed package import documentation.
 
 ## 0.2.0
 
-- Add support for [NIP-02: Contact List and Petnames](https://github.com/nostr-protocol/nips/blob/master/02.md).
+- Added support for [NIP-02: Contact List and Petnames](https://github.com/nostr-protocol/nips/blob/master/02.md).
+
+## 0.3.0
+
+- Reworked `Contact` & `Nostr` class constructors to make instantiating these classes more terse.
+- Added a method to publish arbitrary events [Nostr.sendEvent].